Stalonetray position and colour
Changes stalonetray position to be consistent with SHA: c403238

Changes stalonetray colours to be consistent with SHA: 11245b4

Since the battery bar now has differing width based on state,
stalonetray is positioned to the right in the padding space.
Since stalonetray is only used for nmgui growing SW should not cover
battery bar.